// src/app/rptos/seccion-sistema/interfaces/traslado-response.ts
export interface TrasladoResponse {
    id:               number;
    origen:           number;
    destino:          number;
    costo:            string;
    estado:           string;
    idVehiculo:       number;
    nombre:           string;
    apellido:         string;
    cedula:           string;
    idAuth:           number;
    saldo:            string;
    marca:            string;
    color:            string;
    placa:            string;
    anio_fabricacion: number;
    estado_vehiculo:  string;
    estado_actual:    string;
}